Add configuration option to enable/disable escaping of square brackets
Adds a boolean entry 'escape_square_brackets' to the DEFAULT_OPTS dictionary and allows modification of the default value via a configuration file. Tests the boolean value in _context.text() and only escapes square brackets if this option is set to True.

Escaped square brackets in the source are not unescaped; escaped and unescaped square brackets in the source are preserved in the formatted output when 'escape_square_brackets' is False.