Add configuration option to enable/disable HTML entity substitution
Adds a boolean entry 'entity_substitution' to the DEFAULT_OPTS dictionary and allows modification of the default value via a configuration file. Tests the boolean value in _util.build_mdit() and disables the markdown-it-py 'entity' rule if this option is set to False. Only escapes '&' in _context.text() if 'entity_substitution' is True.